Add test related to weird multilines that are not handled currently.
The specification is unclear on whether it is valid to have
text following the ''', as in:
    desc: '''blabla'''
or
    desc: '''blabla
            hello'''
And even worse, what happens to the whitespaces:
    desc: '''   blabla
            hello'''
And what about the whitespace before and after for a single line?
    desc: '''    blabla    '''
The official parser accepts all of them and ignores the spaces before blabla.