Fix handling of quoteless number-like
The parser would choke on something like:
  x: 0x32
because it think 0x32 is a number due to the initial 0. This is
solved by reading until a space and seeing if the resulting string
is a valid number. If not, it is treated as a string. Note that this
still fails to handle cases like:
  x: 10 apples
because the parser will just eat "x: 10" and continue at "apple".